feat: xxljob vue3实现完成

+++ cc-job-admin/src/main/java/com/cc/job/task/model/vo/TaskInfoVO.java
@@ -22,9 +22,9 @@ public class TaskInfoVO implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    private Integer id;
+    private Long id;
     @Schema(description = "执行器主键ID")
-    private Integer jobGroup;
+    private Long jobGroup;
     private String jobDesc;
     private LocalDateTime addTime;
     private LocalDateTime updateTime;
